requirements.txt
pydub
moviepy
srt
openai-whisper